requirements.txt
boto3==1.12.46
botocore==1.15.46
click==7.1.1
docutils==0.15.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.13.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
